Implement a version checking function for hard forks
If a pegnetd node is not updated late for a hard fork, meaning
a hard fork activates at height X, and the node is updated at X+1,
this code will prevent that node from launching. This is because
the node's state is incorrect. Preventing the node from launching
protects the user from an invalid state. A flag option is provided to
ignore this check if needed.